fix video

+++ client/app/question/question.controller.js
@@ -43,6 +43,9 @@ angular.module('puanJaiApp')
     };
 
     $scope.deleteQuestion = function(){
+      if (!$scope.questionOwner()) {
+        return;
+      }
       var r = confirm("ต้องการลบคําถามนี้?")
       if (r === true) {
         $http.delete('api/questions/' + $stateParams.id).then(function(){
@@ -224,7 +227,6 @@ angular.module('puanJaiApp')
           alert('คําตอบต้องมียาวระหว่าง 40 ถึง 6000 อักขระ');
           return;
         }
-
         var newAnswer = {
           content: $scope.textEditorInput,
           username: $scope.currentUser.username,
